errors with main. repushing
