+++ glow_logic.py
@@ -1,5 +1,5 @@
 import os
-import main
+from api import main
 from slackclient import SlackClient
 from flask import Flask
 from tasks import make_celery
@@ -8,14 +8,18 @@ from config import Config
 app = Flask(__name__)
 app.config.from_object(Config)
 
-b_token = app.config['BOT_TOKEN']
-u_token = app.config['USER_TOKEN']
+# b_token = app.config['BOT_TOKEN']
+# u_token = app.config['USER_TOKEN']
 
 celery = make_celery(app)
 
 
 @celery.task
 def user_iteration(member_list):
+    b_token = os.environ.get('GLOW_BOT_TOKEN')
+
+    sc = SlackClient(b_token)
+    print('step 1')
     members = member_list['members']
     check_list = list(members)
     check_list.remove("UB43F512L")
@@ -29,9 +33,10 @@ def user_iteration(member_list):
 
         user_target_list = list(check_list)
         user_target_list.remove(target)
-        for user in user_target_list:
+        print('step 2')
+        # for user in user_target_list:
 
-            main.sc.api_call("chat.postMessage", channel=target,
-                             text=f"hey <@{target}>! here we glow! Say " +
-                             f"something nice about <@{user}>!",
-                             as_user="true")
+        #     sc.api_call("chat.postMessage", channel=target,
+        #                 text=f"hey <@{target}>! here we glow! Say " +
+        #                 f"something nice about <@{user}>!",
+        #                 as_user="true")
